[#58] feat: BlockContent에서 쓰이는 값들 객체로 관리
