--- frontend/src/utils/regex.ts
@@ -1,9 +1,0 @@
-export const regex = {
-  h1: /^#\s[^\s.]*/gm,
-  h2: /^##\s[^\s.]*/gm,
-  h3: /^###\s[^\s.]*/gm,
-  li: /^-\s[^\s.]*/gm,
-  ol: /^1.\s[^\s.]*/gm,
-  toggle: /^>\s[^\s.]*/gm,
-  quote: /^\|\s[^\s.]*/gm,
-};
